+++ src/main.sh
@@ -9,6 +9,8 @@ function main::exec_tests() {
     test_files+=("$line")
   done < <(helper::load_test_files "$filter" "${files[@]}")
 
+  internal_log "debug" "exec_tests" "filter:$filter" "files:${test_files[*]}"
+
   if [[ ${#test_files[@]} -eq 0 || -z "${test_files[0]}" ]]; then
     printf "%sError: At least one file path is required.%s\n" "${_COLOR_FAILED}" "${_COLOR_DEFAULT}"
     console_header::print_help
@@ -80,6 +82,8 @@ function main::exec_benchmarks() {
   while IFS= read -r line; do
     bench_files+=("$line")
   done < <(helper::load_bench_files "$filter" "${files[@]}")
+
+  internal_log "debug" "exec_benchmarks" "filter:$filter" "files:${bench_files[*]}"
 
   if [[ ${#bench_files[@]} -eq 0 || -z "${bench_files[0]}" ]]; then
     printf "%sError: At least one file path is required.%s\n" "${_COLOR_FAILED}" "${_COLOR_DEFAULT}"